Break out LoggingInputStream for use with any InputStream.
